feat: rename contentscript to content
