+++ package.json
@@ -8,11 +8,11 @@
     "dev": "npm run clear && cross-env NODE_ENV=development run-p dev:*",
     "dev:prepare": "esno scripts/prepare.ts",
     "dev:pages": "vite --config vite.config.pages.ts",
-    "dev:contentScripts": "npm run build:contentScripts --watch src/contentScripts",
-    "build": "cross-env NODE_ENV=production run-s clear build:pages build:prepare build:contentScripts",
+    "dev:content": "npm run build:content --watch src/content",
+    "build": "cross-env NODE_ENV=production run-s clear build:pages build:prepare build:content",
     "build:prepare": "esno scripts/prepare.ts",
     "build:pages": "vite build --config vite.config.pages.ts",
-    "build:contentScripts": "vite build --config vite.config.contentScripts.ts",
+    "build:content": "vite build --config vite.config.content.ts",
     "clear": "rimraf extension/dist extension/manifest.json",
     "lint": "eslint '**/*.{json,ts,js,vue}'"
   },
